distributed code to multiple files/folders to improve readability
