+++ src/consts.ts
@@ -1,0 +1,8 @@
+export const APP_NAME = "yingting's game";
+export const gameName = "Sticker Sketchpad";
+
+export const stickers = [
+    { label: "🐱", icon: "🐱" },
+    { label: "🍎", icon: "🍎" },
+    { label: "🍀", icon: "🍀" },
+];
