+++ src/models/drawable.ts
@@ -1,0 +1,3 @@
+export interface Drawable {
+    display(ctx: CanvasRenderingContext2D): void;
+}
